// lib/types/DecodedInvoice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Description } from "./Description";
import type { Fallback } from "./Fallback";
import type { Features } from "./Features";
import type { Hash32 } from "./Hash32";
import type { HexBytes } from "./HexBytes";
import type { Network } from "./Network";
import type { Payee } from "./Payee";
import type { RouteHop } from "./RouteHop";

export type DecodedInvoice = { network: Network, amount_msat: string | null, timestamp: number, expiry_secs: number, expiry_is_default: boolean, expires_at: number, description: Description | null, payment_hash: Hash32 | null, payment_secret: Hash32 | null, payee: Payee, min_final_cltv_expiry: number, route_hints: Array<Array<RouteHop>>, fallbacks: Array<Fallback>, features: Features, metadata: HexBytes | null, };